+++ src/matmul.hpp
@@ -8,11 +8,11 @@ public:
     int sliceCount;
     int d0;
     int n;
-    int weights0Length; // Number of floats in weights0
+    long weights0Length; // Number of floats in weights0
 
     MatMulSlice(int sliceCount, int n, int d);
-    int splitWeights(int sliceIndex, float* weights, float* weights0);
-    int mergeOutputs(int sliceIndex, float* output, float* output0);
+    long splitWeights(int sliceIndex, float* weights, float* weights0);
+    long mergeOutputs(int sliceIndex, float* output, float* output0);
 };
 
 #endif
